+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Improvements
 - Add the image converter to the extra requires ([#677](../../pull/677))
+- All file tile sources can take either strings or pathlib.Path values ([#683](../../pull/683))
 
 ## Version 1.8.6